+++ src/server.js
@@ -16,6 +16,7 @@
 
 const path = require('path');
 const fs = require('fs');
+require('dotenv').config();
 const express = require('express');
 const session = require('express-session');
 const FileStore = require('session-file-store')(session);